improved status code error handling and further built out tests

+++ back/server/server.go
@@ -64,11 +64,16 @@ func (h *HCShowCalendarServer) getShow(w http.ResponseWriter, r *http.Request) {
 
 	showId := mux.Vars(r)["id"]
 	//TODO perform input validation
+	if showId == "" {
+		code = 400
+		utils.RespondWithError(w, code, errors.New("no id passed to request").Error())
+		return
+	}
 
 	show, err := h.service.GetShow(showId)
 	//determine what type of error and change code and return according error message
 	if err != nil {
-		code = 400
+		code = 500
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
@@ -98,7 +103,7 @@ func (h *HCShowCalendarServer) getShows(w http.ResponseWriter, r *http.Request) 
 	shows, err := h.service.GetShows(showQueryFilters)
 	//TODO build out something to parse errors and deliver error codes
 	if err != nil {
-		code = 400
+		code = 500
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
@@ -114,18 +119,20 @@ func (h *HCShowCalendarServer) createShow(w http.ResponseWriter, r *http.Request
 
 	reqBody, err := io.ReadAll(r.Body)
 	if err != nil {
+		code = 400
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
 	err = json.Unmarshal(reqBody, &show)
 	if err != nil {
+		code = 400
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
 
 	s, err := h.service.CreateShow(show)
 	if err != nil {
-		code = 400
+		code = 500
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
@@ -135,27 +142,34 @@ func (h *HCShowCalendarServer) createShow(w http.ResponseWriter, r *http.Request
 
 func (h *HCShowCalendarServer) updateShow(w http.ResponseWriter, r *http.Request) {
 	defer r.Body.Close()
-	var code int
+	var code = 200
 	var err error
 	var show models.Show
 
 	reqBody, err := io.ReadAll(r.Body)
 	if err != nil {
+		code = 400
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
 	err = json.Unmarshal(reqBody, &show)
 	if err != nil {
+		code = 400
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
 
 	showId := mux.Vars(r)["id"]
+	if showId == "" {
+		code = 400
+		utils.RespondWithError(w, code, errors.New("no id passed to request").Error())
+		return
+	}
 	//perform input validation
 
 	s, err := h.service.UpdateShow(showId, show)
 	if err != nil {
-		code = 400
+		code = 500
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
@@ -165,15 +179,20 @@ func (h *HCShowCalendarServer) updateShow(w http.ResponseWriter, r *http.Request
 
 func (h *HCShowCalendarServer) deleteShow(w http.ResponseWriter, r *http.Request) {
 	var response interface{}
-	var code int
+	var code = 200
 	var err error
 
 	showId := mux.Vars(r)["id"]
+	if showId == "" {
+		code = 400
+		utils.RespondWithError(w, code, errors.New("no id passed to request").Error())
+		return
+	}
 	//perform input validation
 
 	err = h.service.DeleteShow(showId)
 	if err != nil {
-		code = 400
+		code = 500
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
@@ -183,24 +202,26 @@ func (h *HCShowCalendarServer) deleteShow(w http.ResponseWriter, r *http.Request
 
 func (h *HCShowCalendarServer) createUser(w http.ResponseWriter, r *http.Request) {
 	defer r.Body.Close()
-	var code int
+	var code = 201
 	var err error
 	var user models.User
 
 	reqBody, err := io.ReadAll(r.Body)
 	if err != nil {
+		code = 400
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
 	err = json.Unmarshal(reqBody, &user)
 	if err != nil {
+		code = 400
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
 
 	u, err := h.service.CreateUser(user)
 	if err != nil {
-		code = 400
+		code = 500
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
@@ -210,12 +231,13 @@ func (h *HCShowCalendarServer) createUser(w http.ResponseWriter, r *http.Request
 
 func (h *HCShowCalendarServer) authUser(w http.ResponseWriter, r *http.Request) {
 	defer r.Body.Close()
-	var code int
+	var code = 200
 	var err error
 	var u models.User
 
 	reqBody, err := io.ReadAll(r.Body)
 	if err != nil {
+		code = 400
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
@@ -227,7 +249,7 @@ func (h *HCShowCalendarServer) authUser(w http.ResponseWriter, r *http.Request) 
 	}
 	t, err := h.service.AuthUser(u)
 	if err != nil {
-		code = 400
+		code = 500
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
@@ -235,17 +257,27 @@ func (h *HCShowCalendarServer) authUser(w http.ResponseWriter, r *http.Request) 
 }
 
 func (h *HCShowCalendarServer) getUser(w http.ResponseWriter, r *http.Request) {
-	var code int
+	var code = 200
 	var err error
 
 	userId := mux.Vars(r)["id"]
 	//TODO perform input validation
+	if userId == "" {
+		code = 400
+		utils.RespondWithError(w, code, errors.New("no id passed to request").Error())
+		return
+	}
 
 	user, err := h.service.GetUser(userId)
 	//determine what type of error and change code and return according error message
 	if err != nil {
-		code = 400
+		code = 500
 		utils.RespondWithError(w, code, err.Error())
+		return
+	}
+	if user == nil {
+		code = 404
+		utils.RespondWithError(w, code, errors.New("cannot find user").Error())
 		return
 	}
 
@@ -254,27 +286,34 @@ func (h *HCShowCalendarServer) getUser(w http.ResponseWriter, r *http.Request) {
 
 func (h *HCShowCalendarServer) updateUser(w http.ResponseWriter, r *http.Request) {
 	defer r.Body.Close()
-	var code int
+	var code = 200
 	var err error
 	var user models.User
 
 	reqBody, err := io.ReadAll(r.Body)
 	if err != nil {
+		code = 400
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
 	err = json.Unmarshal(reqBody, &user)
 	if err != nil {
+		code = 400
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
 
 	userId := mux.Vars(r)["id"]
 	//perform input validation
+	if userId == "" {
+		code = 400
+		utils.RespondWithError(w, code, errors.New("no id passed to request").Error())
+		return
+	}
 
 	u, err := h.service.UpdateUser(userId, user)
 	if err != nil {
-		code = 400
+		code = 500
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
@@ -284,15 +323,20 @@ func (h *HCShowCalendarServer) updateUser(w http.ResponseWriter, r *http.Request
 
 func (h *HCShowCalendarServer) deleteUser(w http.ResponseWriter, r *http.Request) {
 	var response interface{}
-	var code int
+	var code = 200
 	var err error
 
 	userId := mux.Vars(r)["id"]
 	//perform input validation
+	if userId == "" {
+		code = 400
+		utils.RespondWithError(w, code, errors.New("no id passed to request").Error())
+		return
+	}
 
 	err = h.service.DeleteUser(userId)
 	if err != nil {
-		code = 400
+		code = 500
 		utils.RespondWithError(w, code, err.Error())
 		return
 	}
